requirements.txt
Flask==2.1.1
Pillow==9.0.1
exifread==2.3.2
numpy==1.23.1
opencv-python==4.7.4
matplotlib==3.6.0
scipy==1.8.0
PyMuPDF==1.20.1
PyPDF2==1.26.0